refactor: 디렉터리 구조 유형별로 분류, INIT 삭제
