+++ README.md
@@ -49,8 +49,8 @@
 - [x] 엔터 누르면 금액 사라지는 버그 수정
 - [x] alert 함수 제거
 - [x] 프로젝트 구조 변경
-- [ ] INIT 지우기
-- [ ] 관련된 것들끼리 모으는 식으로 구조 변경
+- [x] INIT 지우기
+- [x] 관련된 것들끼리 모으는 식으로 구조 변경
 
 <br>
