+++ src/js/constants/error-message.js
@@ -1,0 +1,8 @@
+const ERROR_MESSAGE = {
+  PRICE_CANNOT_BE_FLOAT: "금액은 소수가 될 수 없습니다.",
+  PRICE_CANNOT_BE_NEGATIVE: "금액은 자연수여야 합니다.",
+  PRICE_CANNOT_BE_LESS_THAN_THOUSAND: "최소 입력금액은 1000원입니다.",
+  PRICE_CANNOT_BE_OVERWRITTEN: "금액을 중복해서 입력할 수 없습니다.",
+};
+
+export default ERROR_MESSAGE;
