+++ src/js/constants/index.js
@@ -1,0 +1,3 @@
+export { default as ERROR_MESSAGE } from "./error-message.js";
+export { default as LOTTO } from "./lotto.js";
+export { default as SELECTOR } from "./selector.js";
