+++ src/js/constants/lotto.js
@@ -1,0 +1,8 @@
+const LOTTO = {
+  PRICE: 1000,
+  SIZE: 6,
+  MIN_NUM: 1,
+  MAX_NUM: 45,
+};
+
+export default LOTTO;
