+++ src/js/constants/selector.js
@@ -1,0 +1,12 @@
+const SELECTOR = {
+  BUY: "#buy",
+  BUY_INPUT: "#buy-input",
+  BUY_BUTTON: "#buy-button",
+  POCKET: "#pocket",
+  POCKET_TOGGLE: "#pocket-toggle-number",
+  POCKET_LOTTOS: "#pocket-lottos",
+  WINNING: "#winning",
+  POCKET_LOTTO: ".pocket-lotto-numbers",
+};
+
+export default SELECTOR;
