+++ src/js/models/lotto-model.js
