+++ src/js/templates/lotto-template.js
